# internal/httpserver/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Relicta Dashboard API",
    "description": "REST API for the Relicta governance dashboard. Provides release management, governance decisions, actor metrics, approval workflows, and audit trails.",
    "version": "1.0.0",
    "contact": {
      "name": "Relicta",
      "url": "https://github.com/relicta-tech/relicta"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/approvals/pending": {
      "get": {
        "summary": "List pending approvals",
        "operationId": "listPendingApprovals",
        "tags": ["Approvals"],
        "parameters": [
          { "$ref": "#/components/parameters/limit" },
          { "$ref": "#/components/parameters/cursor" },
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/pageSize" },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort field. Prefix with '-' for descending.",
            "schema": {
              "type": "string",
              "enum": ["risk", "-risk", "submitted", "-submitted"],
              "default": "-risk"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of pending approvals",
            "headers": {
              "Link": {
                "description": "RFC 8288 pagination links",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedApprovals"
                }
              }
            }
          }
        }
      }
    },
    "/approvals/{id}/approve": {
      "post": {
        "summary": "Approve a pending release",
        "operationId": "approveRelease",
        "tags": ["Approvals"],
        "parameters": [
          { "$ref": "#/components/parameters/releaseId" }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "justification": { "type": "string" },
                  "force": { "type": "boolean", "default": false }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Release approved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "approved": { "type": "boolean" },
                    "run_id": { "type": "string" },
                    "plan_hash": { "type": "string" },
                    "approved_by": { "type": "string" },
                    "version_next": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/approvals/{id}/reject": {
      "post": {
        "summary": "Reject a pending release",
        "operationId": "rejectRelease",
        "tags": ["Approvals"],
        "parameters": [
          { "$ref": "#/components/parameters/releaseId" }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Release rejected",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rejected": { "type": "boolean" },
                    "run_id": { "type": "string" },
                    "reason": { "type": "string" },
                    "rejected_by": { "type": "string" }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Release not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/audit": {
      "get": {
        "summary": "List audit events",
        "operationId": "listAuditEvents",
        "tags": ["Audit"],
        "parameters": [
          { "$ref": "#/components/parameters/limit" },
          { "$ref": "#/components/parameters/cursor" },
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/pageSize" },
          {
            "name": "release_id",
            "in": "query",
            "description": "Filter by release ID",
            "schema": { "type": "string" }
          },
          {
            "name": "event_type",
            "in": "query",
            "description": "Filter by event type (e.g. PLAN, BUMP, APPROVE)",
            "schema": { "type": "string" }
          },
          {
            "name": "actor",
            "in": "query",
            "description": "Filter by actor ID",
            "schema": { "type": "string" }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Start of time range (RFC3339)",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End of time range (RFC3339)",
            "schema": { "type": "string", "format": "date-time" }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of audit events",
            "headers": {
              "Link": {
                "description": "RFC 8288 pagination links",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedAuditEvents"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from /auth/login"
      }
    },
    "parameters": {
      "limit": {
        "name": "limit",
        "in": "query",
        "description": "Maximum number of items per page (1-100)",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "cursor": {
        "name": "cursor",
        "in": "query",
        "description": "Opaque cursor for cursor-based pagination",
        "schema": { "type": "string" }
      },
      "page": {
        "name": "page",
        "in": "query",
        "description": "Page number for legacy offset pagination",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "pageSize": {
        "name": "page_size",
        "in": "query",
        "description": "Page size for legacy offset pagination (1-100)",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "releaseId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Release run ID",
        "schema": { "type": "string" }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "code"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code",
            "enum": [
              "BAD_REQUEST",
              "INVALID_JSON",
              "VALIDATION_FAILED",
              "MISSING_FIELD",
              "UNAUTHORIZED",
              "INVALID_CREDENTIALS",
              "TOKEN_EXPIRED",
              "TOKEN_INVALID",
              "FORBIDDEN",
              "NOT_FOUND",
              "RELEASE_NOT_FOUND",
              "INTERNAL_ERROR",
              "SERVICE_UNAVAILABLE",
              "AUTH_NOT_CONFIGURED"
            ]
          },
          "details": {
            "description": "Additional error details"
          },
          "request_id": {
            "type": "string",
            "description": "Unique request identifier for tracing"
          }
        }
      },
...
      "Approval": {
        "type": "object",
        "required": ["release_id", "version", "risk_score", "risk_level", "requires_review", "submitted_at", "commit_count"],
        "properties": {
          "release_id": { "type": "string" },
          "version": { "type": "string" },
          "risk_score": { "type": "number", "format": "double" },
          "risk_level": { "type": "string", "enum": ["low", "medium", "high"] },
          "requires_review": { "type": "boolean" },
          "review_reason": { "type": "string" },
          "submitted_at": { "type": "string", "format": "date-time" },
          "submitted_by": { "type": "string" },
          "commit_count": { "type": "integer" },
          "changes": { "type": "array", "items": { "type": "string" } }
        }
      },
      "AuditEvent": {
        "type": "object",
        "required": ["id", "type", "timestamp"],
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string" },
          "release_id": { "type": "string" },
          "actor_id": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "data": { "type": "object", "additionalProperties": true }
        }
      },
...
      "CursorPagination": {
        "type": "object",
        "required": ["data", "total", "limit", "has_more"],
        "properties": {
          "total": { "type": "integer" },
          "limit": { "type": "integer" },
          "next_cursor": { "type": "string" },
          "prev_cursor": { "type": "string" },
          "has_more": { "type": "boolean" }
        }
      },
...
      "PaginatedApprovals": {
        "allOf": [
          { "$ref": "#/components/schemas/CursorPagination" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/Approval" }
              }
            }
          }
        ]
      },
      "PaginatedAuditEvents": {
        "allOf": [
          { "$ref": "#/components/schemas/CursorPagination" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/AuditEvent" }
              }
            }
          }
        ]
      }
    }
  }
}